backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "Backend API"
authors = []

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.32.0"}
sqlalchemy = "^2.0.0"
alembic = "^1.13.0"
psycopg2-binary = "^2.9.0"
boto3 = "^1.35.0"
httpx = "^0.27.0"
apscheduler = "^3.10.0"
pydantic-settings = "^2.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
